feat(tests): Added e2e tests.
